feat: use tags on resource groups

+++ cleanup/Remove-Playground.ps1
@@ -1,21 +1,41 @@
-#This script removes all components of the Virtual WAN Playground
+# This script removes all components of the Virtual WAN Playground
 param(
-  $subscriptionID = '',
-  $rgNames = @("contoso-vwan-rg", "contoso-mgmt-rg", "contoso-spoke1-rg", "contoso-onprem-rg"),
-  ## Number and names of RG's will depend on config.
-  ## Can we parse config file or do a tag-based lookup for simpler lookup? (or add explicit cleanup steps)
-  $cleanuptemplate = "$PSScriptRoot/cleanup.json"
+  [Parameter()]
+  [string]
+  $subscriptionID,
+  [Parameter()]
+  [string]
+  $cleanuptemplate = "$PSScriptRoot/cleanup.json",
+  [Parameter()]
+  [hashtable]
+  $tags = @{Application = 'vwan-playground' }
 )
 
-#Select subscription
+# Select subscription
 Select-AzSubscription -SubscriptionId $subscriptionID
 
-#Remove all resources by deploying and emtpy template using Complete mode
-$jobs = foreach ($rg in $rgNames) {
-  New-AzResourceGroupDeployment -Name "cleanup-$rg" -ResourceGroupName $rg -TemplateFile .\cleanup.json -Mode Complete -Force -AsJob
-}
+# Get resource groups that matches tags
+$rgs = Get-AzResourceGroups -Tag $tags
+$rgNames = $rgs | Select-Object -ExpandProperty ResourceGroupName
 
-#Remove all resource groups
-$rgNames | ForEach-Object -Parallel {
-  Remove-AzResourceGroup -Name $_ -Force
+Write-Host "This will delete resources in the following resource groups:"
+Write-Host "`t$($rgNames -join "`n" )"
+
+$confirmation = Read-Host "Are you Sure You Want To Proceed:"
+if ($confirmation -ne 'y') {
+  Write-Host "Exiting..."
+  exit
+}
+elseif ($confirmation -eq 'y') {
+  #Remove all resources by deploying and emtpy template using Complete mode
+  Write-Host "Removing all resources in selected resource groups..."
+  foreach ($rg in $rgNames) {
+    New-AzResourceGroupDeployment -Name "cleanup-$rg" -ResourceGroupName $rg -TemplateFile .\cleanup.json -Mode Complete -Force -AsJob
+  }
+
+  Write-Host "Removing selected resource groups..."
+  #Remove all resource groups
+  $rgNames | ForEach-Object -Parallel {
+    Remove-AzResourceGroup -Name $_ -Force
+  }
 }
